Can now add/remove/edit jobs
-haven't messed with images so if the clock page would bring up a shift with that job it might explode
-need to make it so it removes all entries in the `Employee Jobs` table where the job has a lower pay than the starting pay

Added dialog option before deleting an employee

+++ Los Alamos Timeclock/Manager/Admin/Editemployees.cs
@@ -28,6 +28,8 @@ namespace Los_Alamos_Timeclock.Manager.Admin
             fieldupdate();
             SSN.TextMaskFormat = MaskFormat.ExcludePromptAndLiterals;
             Phone.TextMaskFormat = MaskFormat.ExcludePromptAndLiterals;
+            jobs.DisplayMember = "getname";
+            jobs.ValueMember = "getpay";
             jobs.DataSource = Main.Joblist;
         }
 
@@ -161,15 +163,21 @@ namespace Los_Alamos_Timeclock.Manager.Admin
 
         private void Delete_Click(object sender, EventArgs e)
         {
-            if (ID == Main.ID)
+            DialogResult result = MessageBox.Show("Are you sure you want to delete " + Fname.Text + " " + Mname.Text + " " + Lname.Text + "? All information related to " + Fname.Text + " " + Mname.Text + " " + Lname.Text + " will also be removed.",
+        "Delete Job?", MessageBoxButtons.YesNo);
+
+            if (result == DialogResult.Yes)
             {
-                MessageBox.Show("You cannot delete your own account");
-            }
-            else
-            {
-                Main.maininstance.sqlinsert("DELETE FROM Employee WHERE ID='" + ID + "'");
-                Main.EmployeeList = Main.maininstance.getEmployees();
-                comboBox1.DataSource = Main.EmployeeList;
+                if (ID == Main.ID)
+                {
+                    MessageBox.Show("You cannot delete your own account");
+                }
+                else
+                {
+                    Main.maininstance.sqlinsert("DELETE FROM Employee WHERE ID='" + ID + "'");
+                    Main.EmployeeList = Main.maininstance.getEmployees();
+                    comboBox1.DataSource = Main.EmployeeList;
+                }
             }
 
         }
@@ -180,10 +188,7 @@ namespace Los_Alamos_Timeclock.Manager.Admin
             Main.maininstance.sqlreader("SELECT JPay FROM `Employee Jobs` WHERE ID='" + ID + "' AND JID='" + jobs.Text + "'");
             if (!Main.reader.HasRows)
             {
-
-                Main.reader.Close();
-                Main.maininstance.sqlreader("SELECT JSPay FROM `Jobs` WHERE JID='" + jobs.Text + "'");
-                pay.Text = Main.reader["JSPay"].ToString();
+                pay.Text = jobs.SelectedValue.ToString();
             }
             else
             {
